woaa-backend/pyproject.toml
[project]
name = "woaa-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alpaca-py>=0.40.1",
    "fastapi>=0.115.12",
    "jose>=1.0.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "sqlalchemy-orm>=1.2.10",
]
